Rename ExpandReferencedPoolsRefactoring to ExpandImportsRefactoring
Pool dictionaries don't exist in their original Smalltalk-80 form any more.
We just have class variables. The nearest equivalent to a pool dictionary
is a Namespace class containing only class variables. The `SharedPool`
Namespace classes are the class equivalents of most of the old
PoolConstantsDictionary instances.

Many of the methods also need to be renamed, but that will be done in a
separate commit to improve the chances of git recognising the class file
delete/add pair as a rename.