Add "Remove Import" refactoring and many other refactoring improvements
- Reorganise the Change Objects hierarchy significantly and add more unit
tests
- Move remaining refactoring classes in Tools namespace and dev sys
packages, and associated tests, into the Refactory packages. The latter are
now so heavily modified, that keeping the Dolphin additions separate is no
longer useful.
- File out 'mourner' flag in the same way as other special behaviour flags
- Add checkboxes for setting new class type, such as bytes/variable or
abstract, to the Create Subclass dialog. Add more validation to the dialog
so that errors are picked up in advance, rather than later throwing errors
from the class builder.
- The result of the CreateSubclassDialog is now a ClassDetails object that
describes everything about the class to be created, and which can also
describe all the other details of a class that are not currently settable
in this dialog. This is then also used elsewhere in the AddClassRefactoring
and AddClassChange to hold details, and can stand in as a proxy for a class
for source filing purposes too.
- Replace the PoolDictionariesShell (appearing as the 'Pool Dictionaries'
tool in the system launcher) with NamespaceShell, that is basically the
same in terms of presentation. The original functionality to create/delete
pool (now namespaces) and add or remove pool variables (now namespace
variables) is still present, but generally delegated to SmalltalkSystem to
perform using the appropriate refactoring (work still needed for adding
variables though).