Experiment with Windows "Safe Subclassing"
"Safe Subclassing" was added to Windows ages ago, but after the original
Dolphin design for low-level integration with Windows was complete. It
improves on the original Windows "subclassing" mechanism by allowing a
window to be subclassed multiple times keeping the individual window procs
independent of one another. It also allows a cookie or closure value to
be passed to the subclass window procedure, and for the subclass to be
removed and re-added easily through a defined API.

Although this basically works, and is somewhat more elegant than the old
style window subclassing mechanism, it has some issues when combined with
Dolphin dialogs (e.g. message boxes) that start another UI process, but may
do so inside a window proc callback that does not return immediately. Once
the dialog is closed, the old UI process may then perform further UI ops
including calling the original window procedure of a control to perform
default processing of the message that caused the dialog to be opened (i.e.
some user input such as a mouse click). If this is done through
DefSubclassProc then an access violation can occur in the control because
presumably it is not expecting to be passed the message after its internal
state has been updated by some other operation. The same problem does not
occur when invoking the original window procedure with CallWindowProc.
The way in which new UI processes are started in Dolphin is problematic for
various reasons, but is difficult to change. Until that is done, it is
(ironically) not safe to use "Safe Subclassing", and so most of this
experiment will need to be backed out.