Upgrade legacy format package when renaming classes out of Smalltalk
The legacy package source format does not support file-out of classes in
namespaces other than Smalltalk. It is intended for loading packages from
older versions, and (with some limitations) round-tripping packages
between versions. As soon as one of the classes in the package is renamed
into another namespace, however, the owning package needs to be upgraded
to be saveable. It makes sense that the Rename Class refactoring do this
in order to simplify the development process of migrating the classes in
a package into namespaces.