Fix regression in system namespace uninitialization breaking deployment
77647b5d31164af7cbc100a364ed40bab18a438b sends setting classes
imports with `imports:` through the class builder, but puts the method in
a development category that is removed by early stages of application
deployment. This prevents the system namespaces imported by Smalltalk
from being removed from those imports when the namespace is uninitialized
for removal. Unresolvable binding references in the Smalltalk imports
list can cause an unbounded recursion when attempting to resolve class
bindings.