Fix infinite binding recursion if import into Smalltalk not found
Removing one of the Namespace classes imported into Smalltalk without
removing the import reference would cause infinite recursion to occur when
attempting to bind an identifier that searched the imports of Smalltalk.
A new NamespaceImport binding reference class is introduced that is
performs a more restricted search of only locally defined variables of
the importable namespaces on the binding path of the import reference.
Imports are canonicalized by the ClassBuilder so that they are always the
primary full path from the Root namespace to the imported namespace, so
there is never any need to search outside the directly defined static
variables of each importable namespace on the path.

Update boot image to consolidated Process instance variable changes and
so all base Class imports are replaced with NamespaceImport instances.