Fix STB upgrade issues in ListView, ScintillaView, etc, and add tests
Improve the regression testing of all STB upgrades of all existing View
classes, especially those with a history of STB version changes.

This was done primarily to validate changes made to the filer for
Namespaces, but also found and fixed a fiew upgrade issues in existing
version converters, e.g. in ListView.

Ported from Dolphin 8.