Revert deletion of OS.ENUMLOGFONTW and ensure is referenced where used
OS.ENUMLOGFONTW was removed as assumed unused (was unreferenced). It is
implicitly referenced from Canvas>>fontFamilies:do:, but only by name in
a string. The method is modified to ensure there is a real reference to
the class from the literal frame which should prevent mistakenly deleting
the class again in future, and which should also ensure that the class is
retained in a deployed application if needed.

This partially reverts commit 1d9faa41760fe6785e6dbd142f6c9c896f374e20