Add #errno and #category custom attributes to Type Library Analyzer
- Functions returning enum types marked with the errno custom attribute
generate the errno external return type, which will cause the VM's FFI
primitive to fail on non-zero return values. This is typically handled in
the library class by throwing an exception, depending on its #invalidCall:
implementation.
- The category custom attribute can be attached to function definitions in
IDL, and it's string value is used to specify an additional method category
for the function. Windows APIs will often classify functions in the same
library into various areas, and this feature can be used to auto-generate
that same classification into the generated library wrapper class.