Short name of class only needs prefix of short name of its namespace
For a class that cannot be bound by unqualified name, it's shortest name
can be shorter than it's full name if its environment has a shorter name.

For example, Graphics.Gdpiplus.Bitmap has a shortname of Gdiplus.Bitmap
because the shortname of the Graphics.Gdiplus namespace is just Gdiplus.
The Graphics namespace is imported into Smalltalk, so Graphics.Gdiplus
does not need it's full name.