Make external POINTs comparable and able to participate in arithmetic
- Reduce the need to unmarshal external POINTs to Points by implementing
comparison and arithmetic protocols on them.
- Fix the long-standing bug with some comparisons of Points against numbers
returning the wrong result. In order to do this we need to double-dispatch
<= comparison as well as <, because it is not true for Points that
(aNumber <= aPoint) is equivalent to (aPoint < aNumber) not. This also
allows us to move #isNaN back down to Number as we don't need to check
for NaN operands in the Magnitude comparison operators any more.