Fix incorrect FFI string encoding conversions with CP_UTF8 active code page
One Windows 10/11 it is possible to run the system with CP_UTF8 as the
active code page. The VM did not handle this properly, because it detects
the code page as multi-byte (which of course, it is), and switches to
1252 instead. It does this because the legacy AnsiString class never really
supported multi-byte code pages correctly. However, it should use the
actual Windows active code page for all FFI encoding conversions, rather
than the code page chosen for AnsiString.