Experiment with configuring Dolphin to always run with CP_UTF8 code page
I thought it would be interesting to see if there was any value in taking
advantage of the capability of recent versions of Windows (post May 2019)
to always run the application with CP_UTF8 as the active code page,
regardless of system configuration. In theory this might allow for some
simplification and performance benefits, as natively Dolphin strings are
UTF-8. Although Windows itself is natively UTF-16, with recent versions the
'A' APIs are supposed to be able to function correctly with CP_UTF8 such
that full Unicode support is available without using the 'W' APIs. Although
the 'A' APIs are presumably translating back and forth to UTF-16 internally
this may be more efficient than performing such translations in the VM, and
certainly than in the image. It also reduces the proliferation of
Utf16String instances returned by 'W' APIs, e.g. for path manipulation.
The different Unicode String encodings are largely transparent in the
Smalltalk system, but the inevitable mixed encodings of Dolphin's native
UTF-8 strings and the Utf16Strings resulting from FFI operations do cause
a lot of encoding transformations to be necessary, and if the API surface
area all operated at UTF-8, this could be avoided.
The approach taken was to add a KernelLibraryUtf8 subclass of KernelLibrary
and then override various operations to use the 'A' APIs where available.
This required migrating some first level API wrapper functionality from
various callers into KernelLibrary, in order to be able to adapt the
encodings appropriately. The system could then detect CP_UTF8 on startup
and switch the KernelLibrary instance referenced by the Kernel32 static
variable to be an instance of KernelLibraryUtf8 instead. This would provide
an elegant solution for correct operation on both recent Windows versions
with general CP_UTF8 support, and also on older Windows versions.
In practice, however, the UTF-8 support provided by the Windows API
versions is incomplete and has a variety of subtle bugs. Unfortunately this
is consistent with previous experience of rigorously testing Windows APIs.
The experiment demonstrates that the CP_UTF8 support is not suffiently
mature to be able to rely on it instead of using the UTF-16 APIs.
Consequently I'll be backing most of this out, but wanted to capture the
experiment in case things improve in future.
The experiment also helped uncover a number of bugs in the handling of
long paths in particular, the fixes for which will of course be retained.