TestCase>>should:raise:matching: failures are difficult to debug
The OA SUnit extension TestCase>>should:raise:matching: extension is useful
but does not provide a straightforward debugging experience in the case
where an Exception of the expected class is raised, but does not match
the discriminator block. The issue is that it suppresses all instances of
the exception class, so by the time the assertion failure is raised the
error details are lost. The extension would be better implemented to trap
only matching exceptions, and otherwise allow exceptions to be passed as
unhandled so that the test errors, providing a better debugging experience
where the unmatched exception and stack are available for investigation
when the test is run.

Ported from Dolphin 8