Fix failure to rename class references in optimized/static expressions
If a method contains a static/optimized expression that references a class
that is renamed, e.g. foo ^##(Bar fullName), then the class reference is
not rewritten. This is because the bindings in ##() expressions are
resolved as if compiling a class method (i.e. metaclass scope).
StClassScope is the lexical scope used to represent class/metaclass binding
scopes in Dolphin's lexical analyzer, and it caches static variable
bindings that are looked up in it for performance reasons, since the
binding resolution is quite expensive. The refactoring rewrite for stale
references clears down the cache of the class scope for each method that is
a rewrite candidate, but doesn't clear it for the metascope. Mostly this is
not required, but if an instance side method includes a ##() expression
that references the renamed class, then the rewrite gets skipped. This
generally results in a binding error when attempting to perform a rename
class (or change namespace) refactoring. Simple to fix and test.