Rename old-style external integer types to Int<size> form; pt 1 file rename
All the historic WORD, DWORD, etc, types used in the Windows APIs and
inherited into Dolphin are rather ambiguous and unclear. WORD can be
confused with the machine "word" size, which therefore varies depending on
the host system. At the time of the early versions of windows, 16-bit chips
were the norm so WORD was the same as machine word. For backwards
compatibility Microsoft had to keep WORD meaning 16-bits when 32-bit
Windows became a thing. This is confusing. If we use the C-style naming
convention Int32, this is clearer since we don't need to know the history.