TestCase>>should:raise: errors rather than asserts on unexpected exception
Missing change from c013f3692 "TestCase>>should:raise:matching: failures
are difficult to debug".

Debugging is easier if the test stops immediately on an error, rather
than handling the unexpected errors and asserting after unwinding the call
stack. Its not particularly useful to have tests fail an assertion (which
may have only a generic description) when an unexpected error occurs. Far
better for the test to error as it would without the #should:raise:. This
also allows for nesting of #should:raise: and #shouldnt:raise: