Ensure WebView2 completion handler callbacks return S_OK
Relying on the block supplied by client code to return S_OK is unwise as
there is no real way to know this is necessary, and even if known it is
easily overlooked. In fact the VM will return 0 in this case anyway, but
it also logs a debug event trace to warn that the callback returned a
non-integer object.