Parse context for env browser on deployed app should be limited
Opening an environment browser on the content of a deployed application
allows one to see exactly what code remains in the application. This is
even more useful if any compilation errors/issues resulting from over-
removal of code are found and displayed. In order to do this we need to
be able to run the browser semantic analysis against an RBModel.

Also ensure that all refactoring dialogs offer the option to show the
changes, rather than applying them immediately.