Improve Scintilla character classification abstraction
Scintilla has 4 separate API functions for updating character
classifications. One resets the classifications to the defaults. Then there
are individual functions to set word character, punctuation and whitespace.
The word char API resets classifications to defaults first, although not
quite the same set of defaults as the specific API (the byte chars outside
the ASCII range are marked as punctuation, whereas normally they are
defaulted to word chars. Frankly this API is badly designed - the
classifications should not be treated as separate settings, since they
all interact. There should be one API that sets the classifications for
all characters together.

This commit introduces a CharacterClassifications abstraction that provides
a better API with a combined representation for the 3 character groups.