File out all to remove Unclassified class category
Class categories are a legacy feature that is not much used now - many
classes are not categorised, and one has to go out of one's way to
categorise a new class. So it made sense to remove the fileout noise for
unclassified classes, which was done as part of d2b8d987e3b73d05c79fd75cfe7b57c3cc3d4c6c
some time ago. A full file-out should have been done at the time to bring
all the source files up to date, but this was overlooked.