Use ICU from icuuc.dll (assumes Windows 10 version 1703, or later)
The small amount of ICU source code in the VM is removed, and we don't
need to export other ICU functions from the VM since we can add and
External.DynamicLinkLibrary wrapper for the underlying DLL.

Windows 10 1703 was released on 5th April 2017. From May 2019, icu.dll
superseded the two earlier DLLs (icuuc.dll and icuin.dll), bit icu.dll
is not present in the Windows Server 2019 OS used in appveyor at present.
Most people are probably running May 2019 or later, or Windows 11, so
will have icu.dll, but for backwards compatibility icuuc.dll is still
present on these systems to forward to icu.dll. There is a small cost
in the redirect when calling the functions from the image, but it
isn't likely to be significant.