Perform conversions from UTF8/16 to Ansi with ICU decoding
All conversions need to use the same UTF decoding algorithms for
consistency, especially in handling invalid sequences. Also fixes the
issue with characters outside the BMP converting to two Ansi replacement
characters.

We don't expect that translations to ANSI will be able to roundtrip as
the code pages can only represent a subset of the Unicode character set.