Upgrade to Scintilla 5.3.4
No change to interface or new features, but does include the following
change that turned out to be rather useful:

Where a multi-byte character contains multiple styles, display each byte
as a representation. This makes it easier to see and fix lexers that change
styles mid-character...

This bug existed in the Smalltalk styling lexer for invalid character
tokens. Only the first code unit of the UTF-8 source was styled with the
invalid-character style.