Add a dark mode to the WebView2 sample
WebView2 supports a dark mode. When set the favicon from the website may
alter in expectation of a dark mode browser caption bar, so it looks odd
if we don't ensure the caption is dark - there is a Win32 API for this at
least. Support for Win32 dark mode controls is undocumented, although used
by the Windows Explorer, so providing a proper dark mode is a non-trivial
exercise. We can do reasonably well, although the toolbar has some rough
edges.