Fix broken Chunk Browser
The venerable Chunk Browser is still a useful tool for browsing and
recovering changes from the change log, but has become a bit broken in
D8 due to changes in the form of various chunks, e.g. class definitions.

To reduce sensitivity to future changes and make maintenance easier, I've
switched the chunk parsing/matching over to use the Smalltalk Scanner.
This is a bit slower, but is guaranteed to scan the code correctly. It
avoids the need for very fragile string/position matching.

I've actioned the todo to switch the system chunks over to using the
default (TZ/locale independent) timestamp printing produced by the ANSI
compliant DateAndTime class.