Include test resource initialization and notifications in the test run log
Initialization of test resources is taking a long time. Include these in
the log to reveal where the time is going (building the SQL Northwind DB
is the answer).

Notifications can contain useful diagnostic information and reveal tests
that are not working as intended. Add logging of the notifications, and
reduce the volume from the existing test suite. Also fix some issues
revealed by the notifications (e.g. UI.Tests.ResourceIdentifierTest>>