Method category emboldening may not be cleared when select new category
When a method is selected in the class browser method list, its categories
are highlighted in the category tree by emboldening the category names.
Normally when the method selection is removed entirely, all emboldening is
also removed, however this doesn't happen if the method selection is
cleared as a side effect of selecting another category in the tree of which
the currently selected method is not a member.