Add buttons to jump to first last sample in profile browser
I find myself often wanting to jump to first/last sample in the profile
browser, and the user experience for this is not very good.

- The use of a scrollbar doesn't work very well here, as it is
designed for paging, so doesn't allow one to position to the last element
if the page size is > 1, so replace with a slider.
- Use toolbar buttons for first, previous, next, last, replacing previous
push buttons for prev/next, an aesthetic improvement.
- Hide the progress bar when not in use, as mostly it is never active and
just looks odd.
- Simplify the samples presenter a bit by adding an index model than can
be tracked by the sub-presenters, reducing the amount of custom logic.
- Disable the buttons when invalid, rather than beeping when pressed.