Remove unused Dictionary protocol from BrowserEnvironment
The #keys as defined did not work correctly with namespaces, but as
browser environments are not dictionaries it makes sense to remove it
rather than fix it. All of the rest of the Dictionary access methods
(at:, at:ifAbsent:, lookup:, associationAt:) may as well go to. I don't
think these are used anywhere in the RB, or the rest of the dev environment
for that matter.