Compiler should add special selectors to literal frame for reference
Being able to browse references mainly relies on selectors being added to
the literal frames of methods. Although the symbols are not required for
special sends, they should still be added to the literal frame so that
reference browsing works as expected. This includes #+ and #-, which can
be replaced with increment/decrement instructions when adding or subtracting
one.

Literals required for reference only are added to the end of the literal
frame. This avoids needlessly lengthening push/send instructions by
increasing the literal index and potentially crossing a threshold such that
a longer instruction form is required.