Add option to show refactoring changes to method rename prompt
The user is prompted when a method rename will affect multiple implementors.
Although the initial method rename dialog includes a check-box to allow one
to opt to see the proposed changes, the subsequent "warning" that the
impact is wider than one class does not. This is new information and might
cause the user to change their mind about wanting to review the refactoring
changes before they are applied.