Inline expression prompts should offer option to show changes
Whether or not an expression is inlined or assigned to a temporary affects
whether or not an inline operation is a true temporary. Inlining the
expression may result in an operation being performed multiple times,
which will change behaviour if the operation has side effects. I usually
choose to inline, and then inspect the result to see whether to undo or
not. An easier option would be to choose to see the proposed changes, and
then cancel if inlining the expression wasn't the right choice after all.