Rename AbstractSemanticAnalyser->SemanticAnalyser (file rename only)
I'm going to make the base semantic analyser concrete, as a fast analyser
for variable binding only, to help with improving auto-completion of
identifiers and various other context sensitive operations initiated from
a method workspace. Currently the full semantic analyser is run only by
the background parsing operation that runs intermittently and relatively
irregularly so as not to interfere with typing. Methods are parsed for
identifier auto-completion anyway, and we might as well run the fast
analysis necessary to map identifiers to variables. This will allow the
identifier auto-completion to offer only those variables in scope, not all
locals defined in a method. Other context sensitive commands such as
browsing references/definitions also work better when the variable analysis
has been done.