Method workspace identifier auto-completion can be more scope aware
- When assembling the list of identifiers we visit the scopes to find all
the declared args and temps. This should take account of the scope at
the insertion point and not add locals defined for other blocks in the
method.
- Environment Browser should only include static variables that are part
of the environment so that, for example, when browsing the content of a
deployed application if editing takes place the auto-completion offerings
are scoped to the classes/variables that survived into the app. Previous
work limited the message completion lists to messages defined in the app.