Work in progress on debugger release->debug method ip mapping
Over the years a number of new bytecode optimisations have been added to
the compiler, and in some cases these have introduced issues with the
mapping of IPs from normal code to debug code that the debugger does when
translating a stack frame to a debug frame. The problems appear to be a
combination of incorrect management of the text maps in the compiler and
the Debugger's ip remapping not handling some corner cases that are either
new or have become more common. This work isn't finished yet, and hence
there are still two versions of the mapping method. One in use, one under
(incomplete) testing.