ChildrenToSiblingsRefactoring fails to push up methods with custom namespace
Refactorings that move methods need to preserve the package of the original
method in case it is compiled into a package default namespace.

Also fixed a disastrous performance problem in PushUpMethodRefactoring.
The algorithm to check sibling supersends is needlessly quadratic in
relation to the total number of classes in the superclass hierarchy, so
attempting to push up a method to Object takes literally minutes to check
the preconditions. This is due to a recursive function using allSubclasses
where it should be using subclasses, as it already recurses down the
hierarchy.

At which point I noticed that the supposed breadth-first traversal of all
subclasses is not actually a breadth-first traversal. It visits all the
immediate subclasses, but then visits all the subclasses of the first
subclass before visiting any of the second subclass. This is a kind of
hybrid with depth first. There are consequences in fixing this that need
to be worked through, so for the moment just fixing in RBAbstractClass by
reverting it to its original allSubclasses implementation (which is actually
breadth-first).