Upgrade (almost) all view resources to STL 6
Autoplay and some contributions are left untouched. This is to provide
some continuing test coverage of view upgrades, and in the case of the
latter to allow for easier interchange of changes with 7.1. The
contributed packages are marked as "legacy", so will save view resources
in STL 4 anyway, but they would pick up any class format changes if resaved