Prompt to remove single/multiple methods should be separately suppressible
Removing a single method is something most will not require a prompt for.
Removing multiple methods is a less common operation, and more impactful,
so one might want to keep the prompt. Also the option to show refactoring
changes can't be combined with the option to suppress, since the new
message box supports only one checkbox. In this case there is no particular
value in seeing the changes since they are just the method removals one
has asked for and been prompted about.