COM functions should be in message send graph for application deployment
Following recent changes to image deployment COM functions required by
implementations of COM interfaces are being removed, with this breaking
deployment of the Simple Web Browser example. This is because the
defineFunctions methods are now being removed. The function tables need to
be considered as sources of message sends, except where the interface is
for egress only, i.e. unless only used for call out. By default all
interfaces are considered potentially implementable, with an opt-in flag
to mark those known to be egress only.