Replace more sends of deprecated ByteArray>>asString
ByteArray>>asString has unpredictable results because it doesn't specify
the encoding. Historically it was assumed to be ANSI code page, but the
results of that would vary depending on the actual system code page. Today
the prevalent string type is UTF-8, but changing the default assumption
of #asString would be too much of a break, so better that it is no longer
used.