VariableBindings should be resolved to live variable on deserialisation
VariableBindings for static variables should be resolved to the live
variable, if present, on deserialization.

For backwards compatibility, old filers should file out as either an
Association or a VariableBinding, and make a best effort to map these to a
live variable on loading, although this is only really possible with class
bindings, and not when loading binary object streams from Dolphin versions
prior to 7.1 where the variables were implemented using the basic
Association class.